Detect FluidSynth >= 2.1.0 during configure step
The integration with FluidSynth will be turned on by default; if version
2.1.0 or newer is missing, then configure script exits with an error.

User can explicitly disable the integration by passing
--disable-fluidsynth parameter.

New feature define is named C_FLUIDSYNTH (in-line with other optional
features) and should be tested in code with 'if C_FLUIDSYNTH'.

Old patches introducing this functionality used to test for the feature
with 'ifdef C_FLUIDSYNTH', which is different than all other features
inside DOSBox.