Silence MSVC 64-bit warning in mem_readd_inline
MSVC reports warning C4244:

  'return': conversion from 'Bitu' to 'Bit32u', possible loss of data

because of virtual function call PageHandler::readd returns Bitu, when
it's expected, that 32-bit value would be returned.

In this particular case, it's ok, that readd returns Bitu - it's a valid
micro-optimization avoiding a number of up/down casting throughout
virtual calls and opening new optimization options for compiler, so we
don't want to remove that (also, it would be quite hard due to all the
classes inheriting from PageHandler).

Also, we can safely expect that Bitu will convert safely back to 32-bit
value in this case, so simple static_cast should suffice to silence the
compiler warning.

Format this function while we're at it - this function is a hot spot
often appearing in the profiler reports, spreading 'if' statements will
help with readability in e.g. Linux perf reports.