Package the speaker setup statements into a function
This intialization function is added to DSP_Reset(), which is
called both during the object's construction as well as requested
on-the-fly by the emulated software, typically during the
detection phase; sometimes many times as the driver performs
various tests to detemine if the card exists and is useable.

This intialization function also makes use of the new DMA
initialization suppression function to ensure "try-and-see
transfers" are not heard.