Differentiate and clarify speaker "volume" as voltage extents
SPKR_VOLUME implies a state, however it's a #define that never
changes, and infact represent a fixed volume, or votage, both
positive and negative.  We therefore rename it to accordingly
given the speaker is pulled from positive to negative.

This is a non-functional change, but makes way for introducing
a neutral voltage.